putting image validation in a util
